refactor: selection concern explicit usage
