+++ src/routes/dashboard/RedoButton/RedoButton.tsx
@@ -1,5 +1,5 @@
 import { Button } from '../../../components';
-import { useSelectController, useUndoHistoryStatus } from '../concerns';
+import { useSelectController, useUndoHistoryStatus } from '../concerns/selection';
 
 const RedoButton = ({ className = '' }: { className?: string }): JSX.Element => {
   const { canRedo } = useUndoHistoryStatus();
