--- src/routes/dashboard/concerns/index.ts
@@ -1,1 +1,0 @@
-export * from './selection';
